fix: quick navigation beteen group async issues
